ELF and weird memory issues fixes

+++ src/mem/vmm.h
@@ -4,7 +4,7 @@
 #include "../util/const.h"
 
 void vmm_init(void);
-void *vmm_alloc(u32 pages);
+void *vmm_alloc(u64 pages);
 void vmm_free(void *mem);
 void vmm_log_status(void);
